chore: useState 폴더링 및 라우터 변경
